Synkronisert med grensesnitt-klasser i bidrag-gcp-proxy + noe refaktorering
