--- src/main/kotlin/no/nav/bidrag/grunnlag/consumer/bidraggcpproxy/api/HentAinntektListeResponse.kt
@@ -1,5 +1,0 @@
-package no.nav.bidrag.gcp.proxy.consumer.inntektskomponenten.response
-
-data class HentAinntektListeResponse(
-  val arbeidsInntektMaaned: List<ArbeidsInntektMaaned>?
-)
