+++ src/main/kotlin/no/nav/bidrag/grunnlag/consumer/bidraggcpproxy/api/ainntekt/HentInntektListeResponse.kt
@@ -1,0 +1,5 @@
+package no.nav.bidrag.grunnlag.consumer.bidraggcpproxy.api.ainntekt
+
+data class HentInntektListeResponse(
+  val arbeidsInntektMaaned: List<ArbeidsInntektMaaned>?
+)
