+++ src/main/kotlin/no/nav/bidrag/grunnlag/consumer/bidraggcpproxy/api/ainntekt/Inntektsmottaker.kt
@@ -1,4 +1,4 @@
-package no.nav.bidrag.gcp.proxy.consumer.inntektskomponenten.response
+package no.nav.bidrag.grunnlag.consumer.bidraggcpproxy.api.ainntekt
 
 data class Inntektsmottaker(
   val identifikator: String,
