+++ src/main/kotlin/no/nav/bidrag/grunnlag/consumer/bidraggcpproxy/api/ainntekt/Tilleggsinformasjon.kt
@@ -1,0 +1,5 @@
+package no.nav.bidrag.grunnlag.consumer.bidraggcpproxy.api.ainntekt
+
+data class Tilleggsinformasjon(
+  val kategori: String
+)
