+++ src/test/kotlin/no/nav/bidrag/grunnlag/BidragGrunnlagTestConfig.kt
@@ -7,8 +7,6 @@ import no.nav.bidrag.commons.web.test.HttpHeaderTestRestTemplate
 import no.nav.bidrag.grunnlag.BidragGrunnlagLocal.Companion.LOCAL_PROFILE
 import no.nav.bidrag.grunnlag.BidragGrunnlagTest.Companion.TEST_PROFILE
 import no.nav.security.mock.oauth2.MockOAuth2Server
-import no.nav.security.token.support.spring.api.EnableJwtTokenValidation
-import no.nav.security.token.support.spring.validation.interceptor.JwtTokenHandlerInterceptor
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.web.client.TestRestTemplate
 import org.springframework.context.annotation.Bean
@@ -16,12 +14,12 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
 import org.springframework.http.HttpHeaders
 
-
 @Configuration
 @OpenAPIDefinition(
     info = Info(title = "bidrag-grunnlag", version = "v1"),
     security = [SecurityRequirement(name = "bearer-key")]
 )
+
 @Profile(TEST_PROFILE, LOCAL_PROFILE)
 class BidragGrunnlagTestConfig {
