+++ src/test/kotlin/no/nav/bidrag/grunnlag/controller/GrunnlagspakkeControllerTest.kt
@@ -2,7 +2,6 @@ package no.nav.bidrag.grunnlag.controller
 
 import no.nav.bidrag.commons.ExceptionLogger
 import no.nav.bidrag.commons.web.HttpHeaderRestTemplate
-import no.nav.bidrag.gcp.proxy.consumer.inntektskomponenten.response.HentAinntektListeResponse
 import no.nav.bidrag.grunnlag.BidragGrunnlagTest
 import no.nav.bidrag.grunnlag.BidragGrunnlagTest.Companion.TEST_PROFILE
 import no.nav.bidrag.grunnlag.TestUtil
@@ -15,6 +14,7 @@ import no.nav.bidrag.grunnlag.api.grunnlagspakke.OpprettGrunnlagspakkeRequest
 import no.nav.bidrag.grunnlag.api.grunnlagspakke.OpprettGrunnlagspakkeResponse
 import no.nav.bidrag.grunnlag.api.grunnlagspakke.PersonIdOgPeriodeRequest
 import no.nav.bidrag.grunnlag.consumer.bidraggcpproxy.BidragGcpProxyConsumer
+import no.nav.bidrag.grunnlag.consumer.bidraggcpproxy.api.ainntekt.HentInntektListeResponse
 import no.nav.bidrag.grunnlag.consumer.bidraggcpproxy.api.skatt.HentSkattegrunnlagResponse
 import no.nav.bidrag.grunnlag.consumer.familiebasak.FamilieBaSakConsumer
 import no.nav.bidrag.grunnlag.consumer.familiebasak.api.FamilieBaSakResponse
@@ -33,7 +33,6 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Assertions.fail
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.mockito.Mockito
@@ -89,8 +88,8 @@ class GrunnlagspakkeControllerTest(
 
     val nyGrunnlagspakkeOpprettetResponse = opprettGrunnlagspakke(OpprettGrunnlagspakkeRequest(Formaal.FORSKUDD,  "X123456"))
 
-    Mockito.`when`(restTemplate.exchange(eq("/inntekt/hent"), eq(HttpMethod.POST), any(), any<Class<HentAinntektListeResponse>>())).thenReturn(
-      ResponseEntity(HentAinntektListeResponse(emptyList()), HttpStatus.OK)
+    Mockito.`when`(restTemplate.exchange(eq("/inntekt/hent"), eq(HttpMethod.POST), any(), any<Class<HentInntektListeResponse>>())).thenReturn(
+      ResponseEntity(HentInntektListeResponse(emptyList()), HttpStatus.OK)
     )
 
     Mockito.`when`(restTemplate.exchange(eq("/skattegrunnlag/hent"), eq(HttpMethod.POST), any(), any<Class<HentSkattegrunnlagResponse>>()))
@@ -110,8 +109,8 @@ class GrunnlagspakkeControllerTest(
 
     assertThat(oppdaterGrunnlagspakkeResponse.grunnlagtypeResponsListe.size).isEqualTo(3)
 
-    oppdaterGrunnlagspakkeResponse.grunnlagtypeResponsListe.forEach() { grunnlagstypeResponse ->
-      grunnlagstypeResponse.hentGrunnlagkallResponseListe.forEach() { hentGrunnlagkallResponse ->
+    oppdaterGrunnlagspakkeResponse.grunnlagtypeResponsListe.forEach { grunnlagstypeResponse ->
+      grunnlagstypeResponse.hentGrunnlagkallResponseListe.forEach { hentGrunnlagkallResponse ->
         assertEquals(hentGrunnlagkallResponse.statuskode, HttpStatus.OK.value())
       }
     }
@@ -122,7 +121,7 @@ class GrunnlagspakkeControllerTest(
 
     val nyGrunnlagspakkeOpprettetResponse = opprettGrunnlagspakke(OpprettGrunnlagspakkeRequest(Formaal.FORSKUDD,  "X123456"))
 
-    Mockito.`when`(restTemplate.exchange(eq("/inntekt/hent"), eq(HttpMethod.POST), any(), any<Class<HentAinntektListeResponse>>())).thenThrow(
+    Mockito.`when`(restTemplate.exchange(eq("/inntekt/hent"), eq(HttpMethod.POST), any(), any<Class<HentInntektListeResponse>>())).thenThrow(
       HttpClientErrorException(HttpStatus.NOT_FOUND)
     )
 
@@ -143,8 +142,8 @@ class GrunnlagspakkeControllerTest(
     assertThat(oppdaterGrunnlagspakkeResponse).isNotNull
     assertThat(oppdaterGrunnlagspakkeResponse.grunnlagtypeResponsListe.size).isEqualTo(3)
 
-    oppdaterGrunnlagspakkeResponse.grunnlagtypeResponsListe?.forEach() { grunnlagstypeResponse ->
-      grunnlagstypeResponse.hentGrunnlagkallResponseListe.forEach() { hentGrunnlagkallResponse ->
+    oppdaterGrunnlagspakkeResponse.grunnlagtypeResponsListe.forEach { grunnlagstypeResponse ->
+      grunnlagstypeResponse.hentGrunnlagkallResponseListe.forEach { hentGrunnlagkallResponse ->
         assertEquals(hentGrunnlagkallResponse.statuskode, HttpStatus.NOT_FOUND.value())
       }
     }
@@ -401,7 +400,7 @@ class GrunnlagspakkeControllerTest(
     if (url != null) {
       json = url.readText()
     } else {
-      fail("Klarte ikke å lese fil fra: ${filpath}")
+      fail("Klarte ikke å lese fil fra: $filpath")
     }
     return json
   }
